kitchen-sink: Menu demo
