+++ example-app/src/demos/index.js
@@ -4,3 +4,4 @@ export * from './Installation'
 export * from './Atoms'
 export * from './Button'
 export * from './Card'
+export * from './Menu'
