+++ src/core/UI/base/reset.css
@@ -84,6 +84,7 @@ time,
 mark,
 audio,
 video {
+    box-sizing: border-box;
     margin: 0 /* rtl:ignore */;
     padding: 0 /* rtl:ignore */;
     border: 0 /* rtl:ignore */;
